fixed #BA1290 https://beyond-aion.com/forum/thread/1290-grankerk%C3%B6nig-belohnungskisten-spawnen-im-boden/

+++ commons/src/com/aionemu/commons/utils/Rnd.java
@@ -46,7 +46,7 @@ public class Rnd {
 	/**
 	 * @see MTRandom#nextFloat()
 	 */
-	public static double nextFloat() {
+	public static float nextFloat() {
 		return rnd.nextFloat();
 	}
